♻️ move globalObject and urlPolyfill to @datadog/js-core/util
