+++ packages/browser-core/src/browser/cookieAccess.spec.ts
@@ -1,8 +1,8 @@
 import { ONE_MINUTE, dateNow } from '@datadog/js-core/time'
+import { globalObject } from '@datadog/js-core/util'
 import type { Clock } from '../../test'
 import { collectAsyncCalls, mockClock, registerCleanupTask, replaceMockable } from '../../test'
 import { display } from '../tools/display'
-import { globalObject } from '../tools/globalObject'
 import { detectVersion, isChromium } from '../tools/utils/browserDetection'
 import type { CookieOptions } from './cookie'
 import { deleteCookie, getCookie, setCookie } from './cookie'
