+++ packages/browser-core/src/browser/cookieAccess.ts
@@ -1,11 +1,11 @@
 import { ONE_MINUTE, ONE_SECOND, dateNow } from '@datadog/js-core/time'
+import { globalObject } from '@datadog/js-core/util'
 import { setInterval, clearInterval } from '../tools/timer'
 import { Observable } from '../tools/observable'
 import { mockable } from '../tools/mockable'
 import { display } from '../tools/display'
 import { generateUUID } from '../tools/utils/stringUtils'
 import { addTelemetryDebug } from '../domain/telemetry'
-import { globalObject } from '../tools/globalObject'
 import { addEventListener, DOM_EVENT, isEventSupported } from './addEventListener'
 import { getCookies, setCookie } from './cookie'
 import type { CookieOptions } from './cookie'
