+++ packages/browser-core/src/browser/fetchObservable.ts
@@ -1,12 +1,11 @@
 import type { ClocksState } from '@datadog/js-core/time'
 import { clocksNow } from '@datadog/js-core/time'
+import { normalizeUrl, globalObject  } from '@datadog/js-core/util'
+import type { GlobalObject } from '@datadog/js-core/util'
 import type { InstrumentedMethodCall } from '../tools/instrumentMethod'
 import { instrumentMethod } from '../tools/instrumentMethod'
 import { monitorError } from '../tools/monitor'
 import { Observable } from '../tools/observable'
-import { normalizeUrl } from '../tools/utils/urlPolyfill'
-import type { GlobalObject } from '../tools/globalObject'
-import { globalObject } from '../tools/globalObject'
 import { readBytesFromStream } from '../tools/readBytesFromStream'
 import { tryToClone } from '../tools/utils/responseUtils'
