+++ packages/browser-core/src/browser/xhrObservable.ts
@@ -1,10 +1,9 @@
 import { elapsed, timeStampNow, clocksNow } from '@datadog/js-core/time'
 import type { Duration, ClocksState } from '@datadog/js-core/time'
+import { normalizeUrl, globalObject  } from '@datadog/js-core/util'
 import type { InstrumentedMethodCall } from '../tools/instrumentMethod'
 import { instrumentMethod } from '../tools/instrumentMethod'
 import { Observable } from '../tools/observable'
-import { normalizeUrl } from '../tools/utils/urlPolyfill'
-import { globalObject } from '../tools/globalObject'
 import { addEventListener } from './addEventListener'
 
 export interface XhrOpenContext {
