+++ packages/browser-core/src/domain/error/trackRuntimeError.ts
@@ -1,9 +1,9 @@
 import { clocksNow } from '@datadog/js-core/time'
+import { globalObject } from '@datadog/js-core/util'
 import { instrumentMethod } from '../../tools/instrumentMethod'
 import { Observable } from '../../tools/observable'
 import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { computeStackTraceFromOnErrorMessage } from '../../tools/stackTrace/computeStackTrace'
-import { globalObject } from '../../tools/globalObject'
 import { DOM_EVENT, isEventSupported } from '../../browser/addEventListener'
 import { noop } from '../../tools/utils/functionUtils'
 import { computeRawError, isError } from './error'
