+++ packages/browser-core/src/domain/telemetry/telemetry.ts
@@ -1,5 +1,5 @@
 import { clocksNow } from '@datadog/js-core/time'
-import { getDebugMode, combine } from '@datadog/js-core/util'
+import { getDebugMode, combine, globalObject, isWorkerEnvironment  } from '@datadog/js-core/util'
 import type { Hook } from '@datadog/js-core/assembly'
 import type { RecursivePartial } from '@datadog/js-core/util'
 import { DISCARDED } from '@datadog/js-core/assembly'
@@ -22,7 +22,6 @@ import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { computeStackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { getConnectivity } from '../connectivity'
 import { canUseEventBridge, getEventBridge, createBatch } from '../../transport'
-import { globalObject, isWorkerEnvironment } from '../../tools/globalObject'
 import { noop } from '../../tools/utils/functionUtils'
 import type { TelemetryEvent } from './telemetryEvent.types'
 import type {
