+++ packages/browser-core/src/tools/sendToExtension.ts
@@ -1,4 +1,4 @@
-import { globalObject } from './globalObject'
+import { globalObject } from '@datadog/js-core/util'
 
 interface BrowserWindow {
   __ddBrowserSdkExtensionCallback?: (message: unknown) => void
