+++ packages/browser-core/src/tools/utils/urlPolyfill.spec.ts
@@ -1,4 +1,4 @@
-import { buildUrl, getPathName, isValidUrl, normalizeUrl, getPristineWindow } from './urlPolyfill'
+import { buildUrl, getPathName, isValidUrl, normalizeUrl, getPristineWindow } from '@datadog/js-core/util'
 
 describe('normalize url', () => {
   it('should resolve absolute paths', () => {
