+++ packages/browser-rum-core/src/domain/contexts/urlContexts.ts
@@ -1,7 +1,8 @@
 import type { RelativeTime } from '@datadog/js-core/time'
 import type { Observable } from '@datadog/browser-core'
 import { relativeNow } from '@datadog/js-core/time'
-import { SESSION_TIME_OUT_DELAY, createValueHistory, mockable, buildUrl, globalObject } from '@datadog/browser-core'
+import { SESSION_TIME_OUT_DELAY, createValueHistory, mockable, globalObject } from '@datadog/browser-core'
+import { buildUrl } from '@datadog/js-core/util'
 import { DISCARDED } from '@datadog/js-core/assembly'
 import type { LocationChange } from '../../browser/locationChangeObservable'
 import type { LifeCycle } from '../lifeCycle'
