+++ packages/js-core/src/entries/util.ts
@@ -3,3 +3,26 @@ export type { Display } from '../util/display'
 export { setDebugMode, getDebugMode } from '../util/debug'
 export * from '../util/mergeInto'
 export * from '../util/typeUtils'
+export { globalObject, isWorkerEnvironment } from '../util/globalObject'
+export type {
+  GlobalObject,
+  Navigator,
+  NetworkInformation,
+  NetworkInterface,
+  EffectiveType,
+  CookieStoreItem,
+  CookieStore,
+  CookieChangeItem,
+  CookieChangeEvent,
+  CookieStoreEventMap,
+  ProfilerFrame,
+  ProfilerStack,
+  ProfilerSample,
+  ProfilerResource,
+  ProfilerTrace,
+  ProfilerInitOptions,
+  SampleBufferFullEvent,
+  Profiler,
+  ProfilerConstructor,
+} from '../util/globalObject'
+export { normalizeUrl, isValidUrl, getPathName, buildUrl, getPristineWindow } from '../util/urlPolyfill'
